Migrate table, update migration docs
